+++ Runtime/AudioLibrary/AudioNamePair.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using UnityEngine;
+
+namespace HephaestusMobile.Audio.SoundsLibrary
+{
+    [Serializable]
+    public class AudioNamePair
+    {
+        public int key;
+        public AudioClip audioClip;
+    }
+}