+++ Runtime/Manager/AudioManager.cs
@@ -28,27 +28,27 @@ namespace HephaestusMobile.Audio.Manager {
         }
 
         /// <inheritdoc cref="IAudioManager"/>
-        public AudioHandler PlayMusicClip(string soundClipName, bool loopSound = true, bool allowMultiple = true, float volume = 0.5f, float delay = 0)
+        public AudioHandler PlayMusicClip(Enum audioClipKey, bool loopSound = true, bool allowMultiple = true, float volume = 0.5f, float delay = 0)
         {
-            return _audioManagerHandler.PlayMusicClip(soundClipName, loopSound, allowMultiple, volume, delay);
+            return _audioManagerHandler.PlayMusicClip(Convert.ToInt32(audioClipKey), loopSound, allowMultiple, volume, delay);
         }
 
         /// <inheritdoc cref="IAudioManager"/>
-        public AudioHandler PlaySoundClip(string soundClipName, bool loopSound = false, bool allowMultiple = true, float volume = 0.5f, float delay = 0)
+        public AudioHandler PlaySoundClip(Enum soundClipName, bool loopSound = false, bool allowMultiple = true, float volume = 0.5f, float delay = 0)
         {
-            return _audioManagerHandler.PlaySoundClip(soundClipName, loopSound, allowMultiple, volume, delay);
+            return _audioManagerHandler.PlaySoundClip(Convert.ToInt32(soundClipName), loopSound, allowMultiple, volume, delay);
         }
 
         /// <inheritdoc cref="IAudioManager"/>
-        public void StopPlayingMusic(string soundClipName)
+        public void StopPlayingMusic(Enum soundClipName)
         {
-            _audioManagerHandler.StopPlayingMusic(soundClipName);
+            _audioManagerHandler.StopPlayingMusic(Convert.ToInt32(soundClipName));
         }
 
         /// <inheritdoc cref="IAudioManager"/>
-        public void StopPlayingSound(string soundClipName)
+        public void StopPlayingSound(Enum soundClipName)
         {
-            _audioManagerHandler.StopPlayingSound(soundClipName);
+            _audioManagerHandler.StopPlayingSound(Convert.ToInt32(soundClipName));
         }
 
         /// <inheritdoc cref="IAudioManager"/>
@@ -76,15 +76,15 @@ namespace HephaestusMobile.Audio.Manager {
         }
 
         /// <inheritdoc cref="IAudioManager"/>
-        public bool IsMusicPlay(string soundClipName)
+        public bool IsMusicPlay(Enum soundClipName)
         {
-            return _audioManagerHandler.IsMusicPlay(soundClipName);
+            return _audioManagerHandler.IsMusicPlay(Convert.ToInt32(soundClipName));
         }
 
         /// <inheritdoc cref="IAudioManager"/>
-        public bool IsSoundPlay(string soundClipName)
+        public bool IsSoundPlay(Enum soundClipName)
         {
-            return _audioManagerHandler.IsSoundPlay(soundClipName);
+            return _audioManagerHandler.IsSoundPlay(Convert.ToInt32(soundClipName));
         }
     }
 }