+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 # Changelog
 
+## [2.0.0] - 2022-02-18
+### Added
++ Constants for audio entities;
++ Custom editor for Constants;
++ Ability to export audio constants;
+
+### Changed
++ Type for audio keys `IAudioManager`, `AudioManager` and from `string` to `Enum`;
++ 
+### Removed
++ Ability use string keys to play and stop audio clips;
+
 ## [1.0.0] - 2022-06-04
-– Implemented basic initialization.
-– Created entity AudioManagerHandler as MonoBehaviour.
-– Made AudioManager non MonoBehaviour.
++ Implemented basic initialization.
++ Created entity `AudioManagerHandler` as `MonoBehaviour`.
++ Made `AudioManager` non `MonoBehaviour`.
 
 ## [0.0.1] - 2021-02-18
-- Created HephaestusCore Audio as UnityPackage.
++ Created HephaestusCore Audio as UnityPackage.